+++ CHANGES.rst
@@ -9,6 +9,10 @@
 
 - Replace ``domain`` with ``bounding_box``. [#74]
 
+- Added a ``LabelMapper`` model where ``mapper`` is an instance of
+  `~astropy.modeling.core.Model`. [#78]
+
+
 0.7 (2016-12-23)
 ----------------